pyproject.toml
[project]
name = "gitforai"
version = "0.1.0"
description = "Git History Mining for AI Agents - Extract, process, and query Git repository history using vector embeddings"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "GitPython>=3.1.40",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
]

llm = [
    "openai>=1.10.0",
    "anthropic>=0.8.0",
    "tiktoken>=0.5.0",  # For token counting
]

local-embeddings = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]

vectordb = [
    "chromadb>=0.4.18",
]

api = [
    "fastapi>=0.108.0",
    "uvicorn[standard]>=0.25.0",
]

all = [
    "gitforai[dev,llm,local-embeddings,vectordb,api]",
]

[project.scripts]
gitforai = "gitforai.cli:app"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/gitforai --cov-report=html --cov-report=term-missing"
